--- a/cmd/orphan.go
+++ b/cmd/orphan.go
@@ -2,10 +2,15 @@ package cmd
 
 import (
 	"encoding/json"
+	"io"
 	"os"
+	"sort"
 	"strings"
 	"sync"
 	"sync/atomic"
+
+	"github.com/dustin/go-humanize"
+	"github.com/spf13/cobra"
 
 	"github.com/autobrr/tqm/client"
 	"github.com/autobrr/tqm/config"
@@ -13,9 +18,6 @@ import (
 	paths "github.com/autobrr/tqm/pathutils"
 	"github.com/autobrr/tqm/torrentfilemap"
 	"github.com/autobrr/tqm/tracker"
-
-	"github.com/dustin/go-humanize"
-	"github.com/spf13/cobra"
 )
 
 var orphanCmd = &cobra.Command{
@@ -179,28 +181,46 @@ var orphanCmd = &cobra.Command{
 
 		wg.Wait()
 
-		// process folders sequentially, since concurrent folder deletion can be problematic
-		var removedLocalFolders uint32
-
+		orphanFolderPaths := make([]string, 0, len(localFolderPaths))
 		for localPath := range localFolderPaths {
-			if tfm.HasPath(localPath, clientDownloadPathMapping) {
-				continue
+			if !tfm.HasPath(localPath, clientDownloadPathMapping) {
+				orphanFolderPaths = append(orphanFolderPaths, localPath)
 			}
+		}
 
+		// Sort orphan folders by path length (depth) in descending order
+		// This ensures deepest directories are processed first
+		sort.Slice(orphanFolderPaths, func(i, j int) bool {
+			return len(orphanFolderPaths[i]) > len(orphanFolderPaths[j])
+		})
+
+		log.Debugf("Processing %d potential orphan folders, sorted by depth", len(orphanFolderPaths))
+
+		var removedLocalFolders uint32
+		for _, localPath := range orphanFolderPaths {
 			log.Info("-----")
-			log.Infof("Removing orphan: %q", localPath)
+			log.Infof("Checking orphan folder: %q", localPath)
 
-			removed := true
+			removed := false
 
-			if flagDryRun {
-				log.Warn("Dry-run enabled, skipping remove...")
+			empty, err := isDirEmpty(localPath)
+			if err != nil {
+				log.WithError(err).Warnf("Could not check if directory is empty, skipping removal: %q", localPath)
+			} else if !empty {
+				log.Warnf("Orphan directory is not empty, skipping removal: %q", localPath)
 			} else {
-				if err := os.Remove(localPath); err != nil {
-					log.WithError(err).Errorf("Failed removing orphan...")
-					atomic.AddUint32(&atomicRemoveFailures, 1)
-					removed = false
+				log.Infof("Attempting to remove empty orphan directory: %q", localPath)
+				if flagDryRun {
+					log.Warn("Dry-run enabled, skipping remove...")
+					removed = true
 				} else {
-					log.Info("Removed")
+					if err := os.Remove(localPath); err != nil {
+						log.WithError(err).Errorf("Failed removing empty orphan directory...")
+						atomic.AddUint32(&atomicRemoveFailures, 1)
+					} else {
+						log.Info("Removed empty orphan directory")
+						removed = true
+					}
 				}
 			}
 
@@ -218,6 +238,26 @@ var orphanCmd = &cobra.Command{
 			Infof("Removed orphans: %d files, %d folders and %d failures",
 				removedLocalFiles, removedLocalFolders, removeFailures)
 	},
+}
+
+func isDirEmpty(path string) (bool, error) {
+	f, err := os.Open(path)
+	if err != nil {
+		return false, err
+	}
+	defer f.Close()
+
+	// Read exactly one entry. If EOF, the directory is empty.
+	// If we get any entry, it's not empty. Poetry.
+	_, err = f.Readdirnames(1)
+	if err == io.EOF {
+		return true, nil
+	}
+	if err != nil {
+		return false, err
+	}
+
+	return false, nil
 }
 
 // processInBatches processes a map in batches using a worker pool