--- a/torrentfilemap/struct.go
+++ b/torrentfilemap/struct.go
@@ -1,9 +1,12 @@
 package torrentfilemap
 
 import (
+	"sync"
+
 	"github.com/autobrr/tqm/config"
 )
 
 type TorrentFileMap struct {
 	torrentFileMap map[string]map[string]config.Torrent
+	pathCache      sync.Map
 }