--- a/torrentfilemap/struct.go
+++ b/torrentfilemap/struct.go
@@ -9,4 +9,5 @@ import (
 type TorrentFileMap struct {
 	torrentFileMap map[string]map[string]config.Torrent
 	pathCache      sync.Map
+	mu             sync.RWMutex
 }