Speeds up the `orphan` scan substantially. Implements fastwalk for faster directory traversal and path caching using sync.Map.

I've also added a grace period for files that were modified within the last 10 minutes. To avoid racing conditions with newly added torrents. Can be customized in config.

- Benchmark 1 is main branch
- Benchmark 2 is https://github.com/autobrr/tqm/pull/17/commits/38e5b9d74a36da175ac64d1b420350b82204b1ab
- Benchmark 3 is https://github.com/autobrr/tqm/pull/17/commits/9a7ce5261b0b0d90bbff0987c2d9205bfa600625 (using https://github.com/charlievieth/fastwalk)

```bash
hyperfine --warmup 1 --runs 5 \
  --setup "sudo sync && sudo sh -c 'echo 3 > /proc/sys/vm/drop_caches'" \
  --prepare "sudo sync && sudo sh -c 'echo 3 > /proc/sys/vm/drop_caches' && mkdir -p /home/user/torrents/qbittorrent/testdata/ && touch /home/user/torrents/qbittorrent/testdata/testfile_{1..100}.tmp" \
  './main orphan qbt --config config.yaml' \
  './no-fastwalk orphan qbt --config config.yaml' \
  './fastwalk orphan qbt --config config.yaml'
```

```
Benchmark 1: ./main orphan qbt --config config.yaml
  Time (mean ± σ):     92.481 s ±  0.767 s    [User: 66.021 s, System: 1.058 s]
  Range (min … max):   91.418 s … 93.535 s    5 runs

Benchmark 2: ./no-fastwalk orphan qbt --config config.yaml
  Time (mean ± σ):     35.109 s ±  3.629 s    [User: 79.196 s, System: 1.315 s]
  Range (min … max):   32.894 s … 41.532 s    5 runs

Benchmark 3: ./fastwalk orphan qbt --config config.yaml
  Time (mean ± σ):     26.861 s ±  0.512 s    [User: 79.098 s, System: 1.386 s]
  Range (min … max):   26.241 s … 27.427 s    5 runs

Summary
  ./fastwalk orphan qbt --config config.yaml ran
    1.31 ± 0.14 times faster than ./no-fastwalk orphan qbt --config config.yaml
    3.44 ± 0.07 times faster than ./main orphan qbt --config config.yaml
```